+++ src/main/java/com/yuuko/core/scheduler/tasks/UpdateMetricsTask.java
@@ -18,7 +18,6 @@ public class UpdateMetricsTask implements Task {
             }
             MetricsManager.getSystemMetrics().update();
             MetricsManager.getAudioMetrics().update();
-            MetricsManager.getCacheMetrics().update();
             MetricsManager.DatabaseInterface.updateMetrics();
         }
     }